requirements.txt
fastapi
uvicorn[standard]
discord.py
aiohttp
aiosqlite
python-dotenv
